Refactoring blood_pressure.dart
